lsp: fix error handling bug
Previously, when an error occurred, handleMessage: would call error: directly,
which caused the thread reading the messages to stop (when the error was abandoned).
This resulted in following blocking requests hanging the image, because no messages were
read anymore. Now, the error: call is done on the UI thread, keeping the message thread running.